result_action: Add Nested Lang support
The ResultAction and the various actions for result
is updated to support Nested Language mode. The API
calls have been updates with an extra argument called
as `nl_file_dict`. It is to this dicitionary that the
result actions will be applying the changes that the
user decides to apply